requirements.txt
flask
flask-sqlalchemy
pyjwt
gunicorn